AI/requirements.txt
dotenv
google-generativeai
pymupdf
fastapi
uvicorn